backend/requirements.txt
boto3==1.17.8
Django==3.1.5
django-cleanup==5.1.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-graphql-jwt==0.3.1
django-markdownx==3.0.1
django-ses==1.0.3
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
Pillow==8.1.0
pip==21.0.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyJWT==1.7.1
setuptools==49.2.1
six==1.15.0
urllib3==1.26.4